Task #230243 sonar duplicate issue solve by creating common component
