+++ src/pages/benefits/viewAllBenefit/ViewAllBenefits.tsx
@@ -27,6 +27,7 @@ import DatePicker from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import { format } from "date-fns";
 import { useNavigate } from "react-router-dom";
+import PaginationList from "./PaginationList";
 const columns = [
   { key: "name", title: "Name", dataType: DataType.String },
   { key: "applicants", title: "Applicants", dataType: DataType.Number },
@@ -75,65 +76,6 @@ const ActionCell = ({ rowData }: ICellTextProps) => {
         size="lg"
       />
     </HStack>
-  );
-};
-const PaginationControls: React.FC<{
-  total: number;
-  pageSize: number;
-  currentPage: number;
-  onPageChange: (page: number) => void;
-}> = ({ total, pageSize, currentPage, onPageChange }) => {
-  const totalPages = Math.ceil(total / pageSize);
-  const pageLimit = 3;
-  const startPage = Math.floor(currentPage / pageLimit) * pageLimit;
-  const endPage = Math.min(startPage + pageLimit, totalPages);
-
-  const handleNext = () => {
-    if (currentPage < totalPages - 1) {
-      onPageChange(currentPage + 1);
-    }
-  };
-
-  const handlePrev = () => {
-    if (currentPage > 0) {
-      onPageChange(currentPage - 1);
-    }
-  };
-
-  return (
-    <Box textAlign="center" mt={4}>
-      <HStack spacing={2} justify="center">
-        {currentPage > 0 && (
-          <Button onClick={handlePrev} colorScheme="blue">
-            Previous
-          </Button>
-        )}
-
-        {Array.from({ length: endPage - startPage }, (_, index) => {
-          const pageIndex = startPage + index;
-          return (
-            <Button
-              key={pageIndex}
-              onClick={() => onPageChange(pageIndex)}
-              colorScheme={currentPage === pageIndex ? "blue" : "gray"}
-              mx={1}
-            >
-              {pageIndex + 1}
-            </Button>
-          );
-        })}
-
-        {currentPage < totalPages - 1 && (
-          <Button
-            onClick={handleNext}
-            colorScheme="blue"
-            aria-label="Go to next page"
-          >
-            Next
-          </Button>
-        )}
-      </HStack>
-    </Box>
   );
 };
 
@@ -364,8 +306,8 @@ const ViewAllBenefits = () => {
             {t("BENEFIT_LIST_TABLE_NO_DATA_MESSAGE")}
           </Text>
         )}
-        <PaginationControls
-          total={data.length}
+        <PaginationList
+          total={data?.length}
           pageSize={pageSize}
           currentPage={pageIndex}
           onPageChange={handlePageChange}
